+++ changelog.md
@@ -16,7 +16,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Changed
 
--
+- added preview to events in admin section
 
 ### Fixed